crypto-multi-agent-system/requirements.txt
# Core Multi-Agent Framework
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.20

# Vector Databases and RAG
chromadb>=0.4.0
# pinecone-client>=3.0.0  # Optional cloud vector DB
sentence-transformers>=2.2.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.1.0
transformers>=4.35.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
aiohttp>=3.8.0
requests>=2.31.0

# Blockchain & Web3
web3>=6.11.0
eth-account>=0.9.0
solana>=0.30.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Time Series & Analytics
influxdb-client>=1.38.0
prophet>=1.1.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
gradio>=4.0.0

# Social Media APIs
tweepy>=4.14.0
python-telegram-bot>=20.6.0
pyrogram>=2.0.0
discord.py>=2.3.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Async & Task Queue
celery>=5.3.0
asyncio-throttle>=1.0.0
schedule>=1.2.0  # Для hourly сканирования

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0
tenacity>=8.2.0  # Retry mechanisms for reliability

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
coverage>=7.3.0

# Security
cryptography>=41.0.0
hashicorp-vault>=0.11.0

# Development
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.5.0

# Optional: For advanced features
# openai>=1.3.0
# anthropic>=0.7.0
